requirements.txt
django
psycopg2
django_rest_framework
